[React] `connect()`ed Components do not respond to new props
The component that results from calling `connect()()` caches the result of mapToProps in its state. This state gets updated when the store's state changes, but not when new props are passed.

This mirrors #152 which fixed the issue in preact's `connect` method. Pretty much the same change here. It uses react's componentWillReceiveProps lifecycle method to compare the old props to the new to see if anything has changed.